fix: refactor code, pay endpoint
